requirements.txt
streamlit>=1.31
python-docx>=1.1.0
pypdf>=4.0.0
reportlab>=4.0.0
pydantic>=2.5
langdetect>=1.0.9
regex>=2023.12.25
spacy>=3.7.2
